automacao-web/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
beautifulsoup4==4.12.2
requests==2.31.0
pandas==2.1.3
openpyxl==3.1.2
python-dotenv==1.0.0
colorama==0.4.6
rich==13.7.0